requirements.txt
numpy; python_version < '3.0'
numpy==1.18.4; python_version >= '3.0'
pygame
matplotlib
opencv-python
Pillow
